+++ src/lib/components/chat/ChatControls.svelte
@@ -13,12 +13,9 @@
 		showEmbeds
 	} from '$lib/stores';
 
-	import Modal from '../common/Modal.svelte';
 	import Controls from './Controls/Controls.svelte';
 	import CallOverlay from './MessageInput/CallOverlay.svelte';
 	import Drawer from '../common/Drawer.svelte';
-	import Overview from './Overview.svelte';
-	import EllipsisVertical from '../icons/EllipsisVertical.svelte';
 	import Artifacts from './Artifacts.svelte';
 	import Embeds from './ChatControls/Embeds.svelte';
 
@@ -154,24 +151,113 @@
 	}
 </script>
 
-<SvelteFlowProvider>
-	{#if !largeScreen}
-		{#if $showControls}
-			<Drawer
-				show={$showControls}
-				onClose={() => {
-					showControls.set(false);
-				}}
+{#if !largeScreen}
+	{#if $showControls}
+		<Drawer
+			show={$showControls}
+			onClose={() => {
+				showControls.set(false);
+			}}
+		>
+			<div
+				class=" {$showCallOverlay || $showOverview || $showArtifacts || $showEmbeds
+					? ' h-screen  w-full'
+					: 'px-4 py-3'} h-full"
 			>
+				{#if $showCallOverlay}
+					<div
+						class=" h-full max-h-[100dvh] bg-white text-gray-700 dark:bg-black dark:text-gray-300 flex justify-center"
+					>
+						<CallOverlay
+							bind:files
+							{submitPrompt}
+							{stopResponse}
+							{modelId}
+							{chatId}
+							{eventTarget}
+							on:close={() => {
+								showControls.set(false);
+							}}
+						/>
+					</div>
+				{:else if $showEmbeds}
+					<Embeds />
+				{:else if $showArtifacts}
+					<Artifacts {history} />
+				{:else if $showOverview}
+					{#await import('./Overview.svelte') then { default: Overview }}
+						<Overview
+							{history}
+							onNodeClick={(e) => {
+								const node = e.node;
+								showMessage(node.data.message, true);
+							}}
+							onClose={() => {
+								showControls.set(false);
+							}}
+						/>
+					{/await}
+				{:else}
+					<Controls
+						on:close={() => {
+							showControls.set(false);
+						}}
+						{models}
+						bind:chatFiles
+						bind:params
+					/>
+				{/if}
+			</div>
+		</Drawer>
+	{/if}
+{:else}
+	<!-- if $showControls -->
+
+	{#if $showControls}
+		<PaneResizer
+			class="relative flex items-center justify-center group border-l border-gray-50 dark:border-gray-850 hover:border-gray-200 dark:hover:border-gray-800  transition z-20"
+			id="controls-resizer"
+		>
+			<div
+				class=" absolute -left-1.5 -right-1.5 -top-0 -bottom-0 z-20 cursor-col-resize bg-transparent"
+			/>
+		</PaneResizer>
+	{/if}
+
+	<Pane
+		bind:pane
+		defaultSize={0}
+		onResize={(size) => {
+			if ($showControls && pane.isExpanded()) {
+				if (size < minSize) {
+					pane.resize(minSize);
+				}
+
+				if (size < minSize) {
+					localStorage.chatControlsSize = 0;
+				} else {
+					// save the size in  pixels to localStorage
+					const container = document.getElementById('chat-container');
+					localStorage.chatControlsSize = Math.floor((size / 100) * container.clientWidth);
+				}
+			}
+		}}
+		onCollapse={() => {
+			showControls.set(false);
+		}}
+		collapsible={true}
+		class=" z-10 bg-white dark:bg-gray-850"
+	>
+		{#if $showControls}
+			<div class="flex max-h-full min-h-full">
 				<div
-					class=" {$showCallOverlay || $showOverview || $showArtifacts || $showEmbeds
-						? ' h-screen  w-full'
-						: 'px-4 py-3'} h-full"
+					class="w-full {($showOverview || $showArtifacts || $showEmbeds) && !$showCallOverlay
+						? ' '
+						: 'px-4 py-3 bg-white dark:shadow-lg dark:bg-gray-850 '} z-40 pointer-events-auto overflow-y-auto scrollbar-hidden"
+					id="controls-container"
 				>
 					{#if $showCallOverlay}
-						<div
-							class=" h-full max-h-[100dvh] bg-white text-gray-700 dark:bg-black dark:text-gray-300 flex justify-center"
-						>
+						<div class="w-full h-full flex justify-center">
 							<CallOverlay
 								bind:files
 								{submitPrompt}
@@ -185,19 +271,28 @@
 							/>
 						</div>
 					{:else if $showEmbeds}
-						<Embeds />
+						<Embeds overlay={dragged} />
 					{:else if $showArtifacts}
-						<Artifacts {history} />
+						<Artifacts {history} overlay={dragged} />
 					{:else if $showOverview}
-						<Overview
-							{history}
-							on:nodeclick={(e) => {
-								showMessage(e.detail.node.data.message);
-							}}
-							on:close={() => {
-								showControls.set(false);
-							}}
-						/>
+						{#await import('./Overview.svelte') then { default: Overview }}
+							<Overview
+								{history}
+								onNodeClick={(e) => {
+									const node = e.node;
+									if (node?.data?.message?.favorite) {
+										history.messages[node.data.message.id].favorite = true;
+									} else {
+										history.messages[node.data.message.id].favorite = null;
+									}
+
+									showMessage(node.data.message, true);
+								}}
+								onClose={() => {
+									showControls.set(false);
+								}}
+							/>
+						{/await}
 					{:else}
 						<Controls
 							on:close={() => {
@@ -209,101 +304,7 @@
 						/>
 					{/if}
 				</div>
-			</Drawer>
+			</div>
 		{/if}
-	{:else}
-		<!-- if $showControls -->
-
-		{#if $showControls}
-			<PaneResizer
-				class="relative flex items-center justify-center group border-l border-gray-50 dark:border-gray-850 hover:border-gray-200 dark:hover:border-gray-800  transition z-20"
-				id="controls-resizer"
-			>
-				<div
-					class=" absolute -left-1.5 -right-1.5 -top-0 -bottom-0 z-20 cursor-col-resize bg-transparent"
-				/>
-			</PaneResizer>
-		{/if}
-
-		<Pane
-			bind:pane
-			defaultSize={0}
-			onResize={(size) => {
-				if ($showControls && pane.isExpanded()) {
-					if (size < minSize) {
-						pane.resize(minSize);
-					}
-
-					if (size < minSize) {
-						localStorage.chatControlsSize = 0;
-					} else {
-						// save the size in  pixels to localStorage
-						const container = document.getElementById('chat-container');
-						localStorage.chatControlsSize = Math.floor((size / 100) * container.clientWidth);
-					}
-				}
-			}}
-			onCollapse={() => {
-				showControls.set(false);
-			}}
-			collapsible={true}
-			class=" z-10 bg-white dark:bg-gray-850"
-		>
-			{#if $showControls}
-				<div class="flex max-h-full min-h-full">
-					<div
-						class="w-full {($showOverview || $showArtifacts || $showEmbeds) && !$showCallOverlay
-							? ' '
-							: 'px-4 py-3 bg-white dark:shadow-lg dark:bg-gray-850 '} z-40 pointer-events-auto overflow-y-auto scrollbar-hidden"
-						id="controls-container"
-					>
-						{#if $showCallOverlay}
-							<div class="w-full h-full flex justify-center">
-								<CallOverlay
-									bind:files
-									{submitPrompt}
-									{stopResponse}
-									{modelId}
-									{chatId}
-									{eventTarget}
-									on:close={() => {
-										showControls.set(false);
-									}}
-								/>
-							</div>
-						{:else if $showEmbeds}
-							<Embeds overlay={dragged} />
-						{:else if $showArtifacts}
-							<Artifacts {history} overlay={dragged} />
-						{:else if $showOverview}
-							<Overview
-								{history}
-								on:nodeclick={(e) => {
-									if (e.detail.node.data.message.favorite) {
-										history.messages[e.detail.node.data.message.id].favorite = true;
-									} else {
-										history.messages[e.detail.node.data.message.id].favorite = null;
-									}
-
-									showMessage(e.detail.node.data.message);
-								}}
-								on:close={() => {
-									showControls.set(false);
-								}}
-							/>
-						{:else}
-							<Controls
-								on:close={() => {
-									showControls.set(false);
-								}}
-								{models}
-								bind:chatFiles
-								bind:params
-							/>
-						{/if}
-					</div>
-				</div>
-			{/if}
-		</Pane>
-	{/if}
-</SvelteFlowProvider>
+	</Pane>
+{/if}
